reorganise server files
